move db connection to __init__.py, add attempt at view page
