+++ flask_app/__init__.py
@@ -3,6 +3,32 @@ The flask application package.
 """
 
 from flask import Flask
+
 app = Flask(__name__)
+
+app.config.from_envvar('APP_SETTINGS', silent=True)
+
+def connect_db():
+    return sqlite3.connect(app.config['DATABASE'])
+
+
+def init_db():
+    with closing(connect_db()) as db:
+        with app.open_resource('flask_app/schema.sql', mode='r') as f:
+            db.cursor().executescript(f.read())
+        db.commit()
+
+
+@app.before_request
+def before_request():
+    g.db = connect_db()
+
+
+@app.teardown_request
+def teardown_request(exception):
+    db = getattr(g, 'db', None)
+    if db is not None:
+        db.close()
+
 
 import flask_app.views
